__init__: fix: PluginManager and server app in the same process
Since PluginManager.init was being called in the main process and the
server was running in a separated process (for being able to restart it
when new plugins were installed), when the server trys to interact with a plugin
that use a separeted thread to run code, they are not executed because treads
attached to the main process were stopped. The solution that I found for
this was to initilize the server and the PluginManager (and eventually
its threads) in the same process.